organizer/requirements.txt
black
Django
dj-database-url
djangorestframework
flake8
psycopg2-binary
Pillow
markdown
pytest-cov
pytest-django
pytest
gunicorn
whitenoise